+++ controllers/tutor/allStudents.js
@@ -1,0 +1,38 @@
+module.exports = function (req,res){
+  let data = {
+    "students": [
+      {
+        "name": "person",
+        "photoURL": "https://via.placeholder.com/100x100",
+        "subjects":[
+          "physics",
+          "math",
+          "english"
+        ],
+        "nextClass": "August 29th, 2018"
+      },
+      {
+        "name": "human",
+        "photoURL": "https://via.placeholder.com/100x100",
+        "subjects":[
+          "chemistry",
+          "math",
+          "french"
+        ],
+        "nextClass": "August 27th, 2018"
+      },
+      {
+        "name": "alien",
+        "photoURL": "https://via.placeholder.com/100x100",
+        "subjects":[
+          "space science",
+          "math",
+          "history of astroids"
+        ],
+        "nextClass": "September 29th, 2018"
+      }
+    ]
+  }
+
+  res.json(data);
+}
