+++ routes/index.js
@@ -1,10 +1,10 @@
 const router = require('express').Router();
 
-const tutorController = require("../controllers");
+const tutorController = require("../controllers/tutor");
 
 module.exports = router
   .get("/next-class", tutorController.nextClass)
   .get("/get-students", tutorController.allStudents)
-  .get("/get-students/:id", tutorController.getOneStudent)
-  .get("/get-students/:id/notes", tutorController.getStudentNotes)
+  .get("/get-students/:id", tutorController.oneStudent)
+  .get("/get-students/:id/notes", tutorController.studentNotes)
